backend/requirements.txt
Werkzeug==2.3.6
Flask==2.3.2
Flask_Cors==3.0.10
Flask_Mail==0.9.1
PyJWT==2.7.0
flask-swagger-ui==4.11.1
pandas==2.0.3
Flask-APScheduler==1.12.4
openai==0.27.8
pytest